requirements.txt
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
facenet-pytorch==2.5.3
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
PyYAML==6.0.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
shellingham==1.5.4
sympy==1.14.0
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.6.1
